Shader => Operation rename, part 3
