+++ packages/fim/src/index.ts
@@ -24,5 +24,5 @@ export { FimDimensions } from './primitives/FimDimensions';
 export { FimError, FimErrorCode } from './primitives/FimError';
 export { FimPoint } from './primitives/FimPoint';
 export { FimRect } from './primitives/FimRect';
-export { FimShaderAlphaBlend } from './shaders/FimShaderAlphaBlend';
-export { FimShaderDarker } from './shaders/FimShaderDarker';
+export { FimOpAlphaBlend } from './ops/FimOpAlphaBlend';
+export { FimOpDarker } from './ops/FimOpDarker';
